nswap: Attempt new detection algorithm
The conclusion from this is that we need "barriers".  The barrier will
be a separation recorded elsewhere between groups of flash moves.  The
determining factor is that when a page is written to again, there must
be a barrier between the two moves.

Practically, there will need to be a barrier between the move of slot 0
up by one, and the swap of the two slots.